LayoutTests/media/track/track-webvtt-tc007-cuenoid-expected.txt

Tests empty cue identifiers (they are optional), but makes sure "-->" found leads to discarded cue.

EXPECTED (video.textTracks[0].cues.length == '3') OK
EXPECTED (video.textTracks[0].cues[0].id == '') OK
EXPECTED (video.textTracks[0].cues[0].startTime == '0.0') OK
EXPECTED (video.textTracks[0].cues[0].endTime == '30.5') OK
EXPECTED (video.textTracks[0].cues[0].getCueAsSource() == 'Bear is Coming!!!!!') OK
EXPECTED (video.textTracks[0].cues[1].id == '') OK
EXPECTED (video.textTracks[0].cues[1].startTime == '31.0') OK
EXPECTED (video.textTracks[0].cues[1].endTime == '60.5') OK
EXPECTED (video.textTracks[0].cues[1].getCueAsSource() == 'I said Bear is coming!!!!') OK
EXPECTED (video.textTracks[0].cues[2].id == '') OK
EXPECTED (video.textTracks[0].cues[2].startTime == '61.0') OK
EXPECTED (video.textTracks[0].cues[2].endTime == '1200.5') OK
EXPECTED (video.textTracks[0].cues[2].getCueAsSource() == 'I said Bear is coming now!!!!') OK
EXPECTED (video.textTracks[1].cues.length == '0') OK
END OF TEST
